add option 'Exit after n iterations'  (J #)
Added an option to exit the script after n iterations.  used the option J for this as not many letters left,  "Jump out of script" maybe?

it just issues a KeyboardInterrupt so we still get the Peak Values line.

added to arg parsing, added to help display, and added script logic.

thanks!